ui: add red default color
