--- Blasphemous.Modding.SkinEditor/Prompts/ColorForm.cs
@@ -1,10 +1,0 @@
-﻿
-namespace Blasphemous.Modding.SkinEditor.Prompts;
-
-public partial class ColorForm : Form
-{
-    public ColorForm()
-    {
-        InitializeComponent();
-    }
-}
