+++ Blasphemous.Modding.SkinEditor/Prompts/ColorPrompt.resx
